requirements.txt

streamlit==1.28.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
pandas==2.1.1
numpy==1.25.2
plotly==5.17.0
matplotlib==3.8.0
seaborn==0.12.2
bibtexparser==1.4.0
PyPDF2==3.0.1
pdfplumber==0.9.0
python-multipart==0.0.6
requests==2.31.0
openai==1.3.0
anthropic==0.3.11
python-dotenv==1.0.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
streamlit-authenticator==0.2.3
json5==0.9.14
